refactor(ts): move kernel/syntax/io/hosts modules into subdirectories
Mechanically relocate sidecar modules into kernel/, syntax/, io/, and
hosts/ subdirectories (each with its co-located tests) and retarget every
#sidecar/<name> specifier to #sidecar/<subdir>/<name>. Delete types.ts:
its Edit type moves into syntax/edits.ts and its Node re-export is dropped
in favour of importing Node directly from #sidecar/syntax/node-schema.

No class-shape, logic, or exported-symbol changes.
