+++ packages/ts/sidecar/src/body-wrapper.ts
@@ -1,8 +1,9 @@
-import { Ast } from '#sidecar/ast';
-import { isErr } from '#sidecar/result';
-import { SourceText } from '#sidecar/source-text';
-import { Sources } from '#sidecar/sources';
-import type { Edit, Node } from '#sidecar/types';
+import { Ast } from '#sidecar/syntax/ast';
+import { isErr } from '#sidecar/kernel/result';
+import { SourceText } from '#sidecar/syntax/source-text';
+import { Sources } from '#sidecar/syntax/sources';
+import type { Edit } from '#sidecar/syntax/edits';
+import type { Node } from '#sidecar/syntax/node-schema';
 
 const STATEMENT_BODY_KEYS: Record<string, string[]> = {
 	DoWhileStatement: ['body'],
