+++ packages/ts/sidecar/src/declaration-reorder.ts
@@ -1,9 +1,9 @@
-import { Ast } from '#sidecar/ast';
-import { Node } from '#sidecar/node-schema';
-import { isErr } from '#sidecar/result';
-import { SourceText } from '#sidecar/source-text';
-import { Sources } from '#sidecar/sources';
-import type { Edit } from '#sidecar/types';
+import { Ast } from '#sidecar/syntax/ast';
+import { Node } from '#sidecar/syntax/node-schema';
+import { isErr } from '#sidecar/kernel/result';
+import { SourceText } from '#sidecar/syntax/source-text';
+import { Sources } from '#sidecar/syntax/sources';
+import type { Edit } from '#sidecar/syntax/edits';
 
 /** Reorders declarations only where the transformation is side-effect safe. */
 export class DeclarationReorder {
