+++ packages/ts/sidecar/src/hosts/vue-script.ts
