+++ packages/ts/sidecar/src/kernel/errors.ts
