+++ packages/ts/sidecar/src/kernel/result.ts
