+++ packages/ts/sidecar/src/syntax/edits.ts
@@ -1,4 +1,14 @@
-import type { Edit } from '#sidecar/types';
+/** A source replacement expressed against the original text offsets. */
+export type Edit = {
+	/** The inclusive replacement start. */
+	start: number;
+
+	/** The exclusive replacement end. */
+	end: number;
+
+	/** The text inserted in place of the selected range. */
+	replacement: string;
+};
 
 /** Applies source edits in offset-safe order. */
 export class Edits {
