+++ packages/ts/sidecar/src/syntax/node-schema.ts
