+++ packages/ts/sidecar/src/syntax/sources.test.ts
@@ -1,7 +1,7 @@
 import assert from 'node:assert/strict';
 import { test } from 'node:test';
-import { isErr } from '#sidecar/result';
-import { Sources } from '#sidecar/sources';
+import { isErr } from '#sidecar/kernel/result';
+import { Sources } from '#sidecar/syntax/sources';
 
 test('Sources.parse returns the program and comments for valid source', () => {
 	const parsed = Sources.parse('sample.ts', 'const one = 1; // note\n');
