+++ packages/ts/sidecar/src/syntax/template-spans.ts
@@ -1,5 +1,5 @@
-import { Ast } from '#sidecar/ast';
-import type { Node } from '#sidecar/types';
+import { Ast } from '#sidecar/syntax/ast';
+import type { Node } from '#sidecar/syntax/node-schema';
 
 type Span = readonly [number, number];
